rename: 연쇄이동
