+++ src/main/java/io/woori/account/wooriaccount/WooriAccountApplication.java
@@ -4,8 +4,9 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
+
 @SpringBootApplication
-@EnableJpaAuditing 
+@EnableJpaAuditing
 public class WooriAccountApplication {
 
     public static void main(String[] args) {
