+++ src/main/java/io/woori/account/wooriaccount/service/inter/CustomerService.java
@@ -1,9 +1,9 @@
 package io.woori.account.wooriaccount.service.inter;
 
-import io.woori.account.wooriaccount.dto.user.CustomerUpdateDTO;
-import io.woori.account.wooriaccount.dto.user.LoginRequestDTO;
-import io.woori.account.wooriaccount.dto.user.LoginResponseDTO;
-import io.woori.account.wooriaccount.dto.user.SignUpRequestDTO;
+import io.woori.account.wooriaccount.dto.customer.CustomerUpdateDTO;
+import io.woori.account.wooriaccount.dto.customer.LoginRequestDTO;
+import io.woori.account.wooriaccount.dto.customer.LoginResponseDTO;
+import io.woori.account.wooriaccount.dto.customer.SignUpRequestDTO;
 
 public interface CustomerService {
